Consider the polar decomposition for gradient reorienation in f3d2 rather than the full affine. Minor changes to remove some warnings.

+++ reg-lib/_reg_aladin.h
@@ -121,8 +121,8 @@ class reg_aladin
       SetMacro(LevelsToPerform,unsigned int);
       GetMacro(LevelsToPerform,unsigned int);
 
-      SetMacro(BlockPercentage,float);
-      GetMacro(BlockPercentage,float);
+      SetMacro(BlockPercentage,int);
+      GetMacro(BlockPercentage,int);
 
       SetMacro(InlierLts,float);
       GetMacro(InlierLts,float);
